rename file place to sight
