+++ lib/ui/screen/sight_details_screen/sight_details_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:places/constants/app_strings.dart';
-import 'package:places/data/model/place.dart';
+import 'package:places/data/model/sight.dart';
 import 'package:places/styles/custom_icons.dart';
 import 'package:places/ui/screen/sight_details_screen/sight_details_header_delegate.dart';
 import 'package:places/ui/widgets/icon_text_button.dart';
