+++ lib/ui/screen/sight_list_screen/sight_list_landscape_orientation.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../../data/model/place.dart';
+import '../../../data/model/sight.dart';
 import '../../widgets/sight_cards/sight_card.dart';
 
 class SightListLandscapeOrientation extends StatelessWidget {
