+++ lib/ui/screen/visiting_screen/visited_page.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:places/constants/app_strings.dart';
-import 'package:places/data/model/place.dart';
+import 'package:places/data/model/sight.dart';
 import 'package:places/styles/custom_icons.dart';
 import 'package:places/ui/screen/visiting_screen/draggable_sight_cards_list_view.dart';
 import 'package:places/ui/screen/visiting_screen/empty_list_page.dart';
